requirements.txt
mecab-python3
boto3
pandas